* win/makefile.vc:  Updated MSVC build to properly deal with * win/nmakehlp.c:   MSVC8 and AMD64 target. Backport from 8.5 * win/rules.vc: * generic/tcl.h:    Fixed stat definition for MSVC8 AMD64. * win/tclWinSock.c: Casting type police. * win/tclWinTime.c:

+++ win/tclWinSock.c
@@ -2356,7 +2356,7 @@ SocketThread(LPVOID arg)
 
     SetEvent(tsdPtr->readyEvent);
 
-    return msg.wParam;
+    return (DWORD)msg.wParam;
 }
